requirements.txt
pymongo==4.14.0
python-dotenv==1.1.1
Requests==2.32.4
tqdm==4.67.1
websocket_client==1.8.0
